third_party/WebKit/LayoutTests/editing/selection/deleteFromDocument-expected.txt

PASS span1.textContent is "foobar"
PASS span1.textContent is "foobar"
PASS span1.firstChild.nextSibling.id is "span2"
PASS span1.firstChild.nextSibling.firstChild.nodeValue is ""
PASS test.textContent is "fz"
PASS successfullyParsed is true

TEST COMPLETE

fz
